test(mitm-addon): split response streaming tests
