--- crates/runner/mitm-addon/tests/test_response_streaming.py
@@ -1,1265 +1,0 @@
-"""Tests for response streaming parser setup."""
-
-import gzip
-import zlib
-
-import pytest
-from mitmproxy import http
-from mitmproxy.test import tutils
-
-import flow_metadata_keys as metadata_keys
-import mitm_addon
-import response_streaming
-from body_limits import LARGE_RESPONSE_DECOMPRESS_LIMIT, STREAM_BUFFER_LIMIT
-from tests.flow_helpers import header_map, response_stream
-from tests.jsonl_log_helpers import jsonl_exists_after_flush, read_jsonl_entries_after_flush
-from tests.x_flow_helpers import make_x_pipeline_flow, make_x_response_flow
-
-_OVERSIZED_NDJSON_LINE_BYTES = LARGE_RESPONSE_DECOMPRESS_LIMIT + 1024
-
-
-class TestResponseStreamBuffer:
-    """Tests for generic response stream buffering behavior."""
-
-    def _json_response_flow(
-        self,
-        real_flow,
-        *,
-        host: str = "api.example.com",
-        capture_body: bool = True,
-    ) -> http.HTTPFlow:
-        flow = real_flow(with_response=False, host=host)
-        if capture_body:
-            flow.metadata[metadata_keys.CAPTURE_BODY] = True
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-        return flow
-
-    def test_stream_callback_buffers_chunks(self, real_flow):
-        flow = self._json_response_flow(real_flow)
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        assert callback(b"hello ") == b"hello "
-        assert callback(b"world") == b"world"
-        assert bytes(flow.metadata[metadata_keys.STREAM_BUFFER]) == b"hello world"
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is False
-
-    def test_stream_callback_stops_buffering_at_limit(self, real_flow):
-        flow = self._json_response_flow(real_flow)
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        chunk = b"x" * STREAM_BUFFER_LIMIT
-        assert callback(chunk) == chunk
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is False
-
-        assert callback(b"overflow") == b"overflow"
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-
-    def test_stream_callback_large_single_chunk(self, real_flow):
-        flow = self._json_response_flow(real_flow)
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        big_chunk = b"A" * (STREAM_BUFFER_LIMIT + 1000)
-        assert callback(big_chunk) == big_chunk
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-
-    def test_stream_callback_partial_fill_then_overflow(self, real_flow):
-        flow = self._json_response_flow(real_flow)
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        half = STREAM_BUFFER_LIMIT // 2
-        callback(b"A" * half)
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is False
-
-        callback(b"B" * STREAM_BUFFER_LIMIT)
-        remaining = STREAM_BUFFER_LIMIT - half
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER][:half] == bytearray(b"A" * half)
-        assert flow.metadata[metadata_keys.STREAM_BUFFER][half:] == bytearray(b"B" * remaining)
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-
-    def test_stream_callback_empty_chunk(self, real_flow):
-        flow = self._json_response_flow(real_flow)
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        assert callback(b"") == b""
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == 0
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is False
-
-        callback(b"hello")
-        assert bytes(flow.metadata[metadata_keys.STREAM_BUFFER]) == b"hello"
-
-    def test_non_capture_response_does_not_retain_body_prefix(self, real_flow):
-        flow = self._json_response_flow(
-            real_flow,
-            host="api.github.com",
-            capture_body=False,
-        )
-
-        mitm_addon.responseheaders(flow)
-
-        body = b"x" * (STREAM_BUFFER_LIMIT + 1000)
-        assert response_stream(flow)(body) == body
-
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-        assert flow.metadata[metadata_keys.RESPONSE_STREAM_STATE]["total_bytes"] == len(body)
-
-    def test_billable_connector_without_body_parser_does_not_retain_prefix(self, real_flow):
-        flow = self._json_response_flow(
-            real_flow,
-            host="api.gamma.example",
-            capture_body=False,
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "gamma"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-
-        mitm_addon.responseheaders(flow)
-
-        response_stream(flow)(b"g" * (STREAM_BUFFER_LIMIT + 1000))
-
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-        assert metadata_keys.X_NDJSON_STATE not in flow.metadata
-        assert "connector_response_finish" not in flow.metadata
-
-
-class TestResponseEncodingInspectionRisk:
-    """Tests structured risk signals at the actual response decoder gate."""
-
-    def _model_flow(
-        self,
-        real_flow,
-        tmp_path,
-        *,
-        content_encoding: str,
-        content_type: str = "application/json",
-    ) -> http.HTTPFlow:
-        flow = real_flow(
-            with_response=False,
-            host="api.anthropic.com",
-            path="/v1/messages",
-            method="POST",
-        )
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {
-                    "content-type": content_type,
-                    "content-encoding": content_encoding,
-                }
-            ),
-        )
-        flow.metadata.update(
-            {
-                metadata_keys.VM_RUN_ID: "run-encoding-risk",
-                metadata_keys.VM_PROXY_LOG_PATH: str(tmp_path / "proxy.jsonl"),
-                metadata_keys.FIREWALL_NAME: "model-provider:anthropic-api-key",
-                metadata_keys.FIREWALL_BILLABLE: True,
-                metadata_keys.MODEL_USAGE_PROVIDER: "claude-sonnet-4-6",
-            }
-        )
-        return flow
-
-    @pytest.mark.parametrize(
-        ("content_encoding", "content_type", "decode_skip_reason"),
-        [
-            pytest.param(
-                "zstd",
-                "application/json",
-                "zstd streaming output cannot be hard-bounded",
-                id="model-json-zstd",
-            ),
-            pytest.param(
-                "br",
-                "text/event-stream",
-                "brotli streaming output cannot be bounded",
-                id="model-sse-brotli",
-            ),
-            pytest.param(
-                "private-encoding-value",
-                "application/json",
-                "unsupported content encoding",
-                id="model-json-unknown",
-            ),
-        ],
-    )
-    def test_observable_model_provider_logs_non_streamable_encoding_risk(
-        self,
-        real_flow,
-        tmp_path,
-        mitm_ctx,
-        content_encoding: str,
-        content_type: str,
-        decode_skip_reason: str,
-    ) -> None:
-        flow = self._model_flow(
-            real_flow,
-            tmp_path,
-            content_encoding=content_encoding,
-            content_type=content_type,
-        )
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        [entry] = read_jsonl_entries_after_flush(tmp_path / "proxy.jsonl")
-        assert entry["type"] == "usage_underbilling"
-        assert entry["reason"] == "response_encoding_not_stream_decodable"
-        assert entry["underbilling_class"] == "risk"
-        assert entry["component"] == "mitm_addon"
-        assert entry["run_id"] == "run-encoding-risk"
-        assert entry["firewall_name"] == "model-provider:anthropic-api-key"
-        assert entry["status_code"] == 200
-        assert entry["decode_skip_reason"] == decode_skip_reason
-        assert callable(response_stream(flow))
-
-    def test_unknown_encoding_value_is_not_persisted(self, real_flow, tmp_path, mitm_ctx) -> None:
-        flow = self._model_flow(
-            real_flow,
-            tmp_path,
-            content_encoding="private-encoding-value",
-        )
-
-        with mitm_ctx() as log:
-            mitm_addon.responseheaders(flow)
-
-        [entry] = read_jsonl_entries_after_flush(tmp_path / "proxy.jsonl")
-        assert "private-encoding-value" not in str(entry)
-        assert "private-encoding-value" not in log.debug.call_args.args[0]
-        assert entry["decode_skip_reason"] == "unsupported content encoding"
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    def test_stream_safe_model_response_keeps_parser_without_risk(
-        self, real_flow, tmp_path, mitm_ctx
-    ) -> None:
-        flow = self._model_flow(real_flow, tmp_path, content_encoding="gzip")
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        assert "model_json_usage_finish" in flow.metadata
-        assert not jsonl_exists_after_flush(tmp_path / "proxy.jsonl")
-
-    def test_non_observable_model_response_does_not_log_encoding_risk(
-        self, real_flow, tmp_path, mitm_ctx
-    ) -> None:
-        flow = self._model_flow(real_flow, tmp_path, content_encoding="zstd")
-        flow.metadata.pop(metadata_keys.MODEL_USAGE_PROVIDER)
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        assert not jsonl_exists_after_flush(tmp_path / "proxy.jsonl")
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    @pytest.mark.parametrize(
-        ("request_method", "response_status"),
-        [
-            pytest.param("GET", 101, id="informational"),
-            pytest.param("GET", 204, id="no-content"),
-            pytest.param("GET", 205, id="reset-content"),
-            pytest.param("GET", 304, id="not-modified"),
-            pytest.param("HEAD", 200, id="head"),
-            pytest.param("CONNECT", 200, id="successful-connect"),
-        ],
-    )
-    def test_bodyless_model_response_does_not_log_encoding_risk(
-        self,
-        real_flow,
-        tmp_path,
-        mitm_ctx,
-        request_method: str,
-        response_status: int,
-    ) -> None:
-        flow = self._model_flow(real_flow, tmp_path, content_encoding="zstd")
-        flow.request.method = request_method
-        assert flow.response is not None
-        flow.response.status_code = response_status
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        assert not jsonl_exists_after_flush(tmp_path / "proxy.jsonl")
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    def test_successful_billable_connector_logs_non_streamable_encoding_risk(
-        self, real_flow, tmp_path, mitm_ctx
-    ) -> None:
-        flow = make_x_pipeline_flow(real_flow, tmp_path, content_encoding="zstd")
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        [entry] = read_jsonl_entries_after_flush(tmp_path / "proxy.jsonl")
-        assert entry["type"] == "usage_underbilling"
-        assert entry["reason"] == "response_encoding_not_stream_decodable"
-        assert entry["underbilling_class"] == "risk"
-        assert entry["firewall_name"] == "x"
-        assert entry["status_code"] == 200
-        assert entry["decode_skip_reason"] == "zstd streaming output cannot be hard-bounded"
-
-    def test_unknown_connector_encoding_does_not_retain_unusable_fallback(
-        self, real_flow, tmp_path, mitm_ctx
-    ) -> None:
-        flow = make_x_pipeline_flow(
-            real_flow,
-            tmp_path,
-            content_encoding="private-encoding-value",
-        )
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        [entry] = read_jsonl_entries_after_flush(tmp_path / "proxy.jsonl")
-        assert entry["reason"] == "response_encoding_not_stream_decodable"
-        assert entry["decode_skip_reason"] == "unsupported content encoding"
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    @pytest.mark.parametrize(
-        ("firewall_name", "firewall_billable", "response_status"),
-        [
-            pytest.param("x", False, 200, id="non-billable"),
-            pytest.param("stripe", True, 200, id="no-registered-parser"),
-            pytest.param("x", True, 400, id="unsuccessful-response"),
-            pytest.param("x", True, 204, id="bodyless-response"),
-        ],
-    )
-    def test_non_inspected_connector_does_not_log_encoding_risk(
-        self,
-        real_flow,
-        tmp_path,
-        mitm_ctx,
-        firewall_name: str,
-        firewall_billable: bool,
-        response_status: int,
-    ) -> None:
-        flow = make_x_pipeline_flow(
-            real_flow,
-            tmp_path,
-            response_status=response_status,
-            content_encoding="zstd",
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = firewall_name
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = firewall_billable
-
-        with mitm_ctx():
-            mitm_addon.responseheaders(flow)
-
-        assert not jsonl_exists_after_flush(tmp_path / "proxy.jsonl")
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-
-class TestNdjsonExtractor:
-    """Tests for X NDJSON extraction through responseheaders (issue #9534)."""
-
-    def _stream_flow(self, real_flow, *, capture_body=False):
-        flow = make_x_response_flow(real_flow, path="/2/tweets/search/stream")
-        if capture_body:
-            flow.metadata[metadata_keys.CAPTURE_BODY] = True
-        mitm_addon.responseheaders(flow)
-        return flow
-
-    def _stream_parser(self, real_flow):
-        flow = self._stream_flow(real_flow)
-        return response_stream(flow), flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-    def test_single_line(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n')
-        assert state["data_count"] == 1
-        assert state["includes"] == {"users": 1}
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 0
-
-    def test_multiple_lines_aggregate_counts(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n')
-        parse(b'{"data":{"id":"2"},"includes":{"users":[{"id":"u2"},{"id":"u3"}]}}\n')
-        assert state["data_count"] == 2
-        assert state["includes"] == {"users": 3}
-        assert state["lines_parsed"] == 2
-
-    def test_forensic_buffer_truncation_does_not_stop_parser(self, real_flow):
-        flow = self._stream_flow(real_flow, capture_body=True)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-        parse(b'{"data":{"id":"1"}}\n' + b"x" * (200 * 1024))
-
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-        assert state["data_count"] == 1
-        assert "connector_response_finish" in flow.metadata
-
-    def test_decompresses_gzip_stream_before_parsing(self, real_flow):
-        ndjson_body = (
-            b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n'
-            b'{"data":{"id":"2"},"includes":{"users":[{"id":"u2"}]}}\n'
-            b'{"data":{"id":"3"},"includes":{"users":[{"id":"u3"}]}}\n'
-        )
-        compressed = gzip.compress(ndjson_body)
-        flow = make_x_response_flow(
-            real_flow,
-            path="/2/tweets/search/stream",
-            content_encoding="gzip",
-        )
-
-        mitm_addon.responseheaders(flow)
-
-        mid = len(compressed) // 2
-        response_stream(flow)(compressed[:mid])
-        response_stream(flow)(compressed[mid:])
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-        assert state["data_count"] == 3
-        assert state["includes"] == {"users": 3}
-        assert "connector_response_finish" in flow.metadata
-
-    def test_chunked_line_split_mid_json(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"},"include')
-        parse(b's":{"users":[{"id":"u1"}]}}\n')
-        assert state["data_count"] == 1
-        assert state["includes"] == {"users": 1}
-        assert state["lines_parsed"] == 1
-
-    def test_keep_alive_blank_lines(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b"\n\n")
-        parse(b'{"data":{"id":"1"}}\n')
-        parse(b"\n")
-        parse(b'{"data":{"id":"2"}}\n')
-        assert state["data_count"] == 2
-        assert state["lines_parsed"] == 2
-
-    def test_crlf_line_endings(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"}}\r\n{"data":{"id":"2"}}\r\n')
-        assert state["data_count"] == 2
-        assert state["lines_parsed"] == 2
-
-    def test_crlf_line_ending_split_across_chunks(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"}}\r')
-        parse(b'\n{"data":{"id":"2"}}\r\n')
-        assert state["data_count"] == 2
-        assert state["lines_parsed"] == 2
-        assert state["lines_failed"] == 0
-
-    def test_malformed_line_increments_failures(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"}}\n')
-        parse(b"not json at all\n")
-        parse(b'{"data":{"id":"2"}}\n')
-        assert state["data_count"] == 2
-        assert state["lines_parsed"] == 2
-        assert state["lines_failed"] == 1
-
-    def test_invalid_utf8_line_increments_failures_and_continues(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'\x80{"data":{"id":"bad"}}\n{"data":{"id":"after"}}\n')
-
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 1
-
-    def test_truncated_trailing_line_not_counted(self, real_flow):
-        """Connection drops mid-line — partial trailing line stays in buf, not counted."""
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":{"id":"1"}}\n{"data":{"id":"2"}')  # no trailing \n
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-
-    def test_complete_trailing_line_without_newline_counted_on_finish(self, real_flow):
-        flow = self._stream_flow(real_flow)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}')
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert state["data_count"] == 1
-        assert state["includes"] == {"users": 1}
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 0
-
-    def test_incomplete_trailing_line_without_newline_fails_on_finish(self, real_flow):
-        flow = self._stream_flow(real_flow)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-        parse(b'{"data":{"id":"1"}}\n{"data":{"id":"2"}')
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 1
-
-    @pytest.mark.parametrize("tail", [b"", b"\r", b"\r\r"])
-    def test_blank_trailing_line_ignored_on_finish(self, real_flow, tail):
-        flow = self._stream_flow(real_flow)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-        parse(b'{"data":{"id":"1"}}\n' + tail)
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 0
-
-    def test_trailing_line_finish_is_idempotent(self, real_flow):
-        flow = self._stream_flow(real_flow)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-
-        parse(b'{"data":{"id":"1"}}')
-        response_streaming.finalize_connector_response_state(flow)
-        finalized = dict(state)
-        finalized["includes"] = dict(state["includes"])
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert state == finalized
-
-    def test_oversized_line_tail_not_counted_on_finish(self, real_flow):
-        flow = self._stream_flow(real_flow)
-        parse = response_stream(flow)
-        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
-        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
-
-        parse(big)
-        parse(b'{"data":{"id":"tail"}}')
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert state["data_count"] == 0
-        assert state["lines_parsed"] == 0
-        assert state["lines_failed"] == 1
-
-    def test_empty_chunks_safe(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b"")
-        parse(b'{"data":{"id":"1"}}\n')
-        parse(b"")
-        assert state["data_count"] == 1
-
-    def test_oversized_line_dropped(self, real_flow):
-        """Oversized line is dropped; subsequent lines parse normally."""
-        parse, state = self._stream_parser(real_flow)
-        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
-        parse(big)
-        parse(b"\n")
-        parse(b'{"data":{"id":"after"}}\n')
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 1
-
-    def test_oversized_line_discards_until_newline(self, real_flow):
-        """A valid-looking tail of an overlong line must not be counted as its own row."""
-        parse, state = self._stream_parser(real_flow)
-        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
-        parse(big)
-        parse(b'{"data":{"id":"tail"}}\n')
-        parse(b'{"data":{"id":"next"}}\n')
-
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 1
-
-    def test_oversized_line_with_newline_continues_in_same_chunk(self, real_flow):
-        """Dropping an overlong row should not discard valid later rows in the same chunk."""
-        parse, state = self._stream_parser(real_flow)
-        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
-        parse(big + b'\n{"data":{"id":"after"}}\n')
-
-        assert state["data_count"] == 1
-        assert state["lines_parsed"] == 1
-        assert state["lines_failed"] == 1
-
-    def test_includes_multiple_keys(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(
-            b'{"data":{"id":"1"},"includes":'
-            b'{"users":[{"id":"u1"}],'
-            b'"tweets":[{"id":"t1"},{"id":"t2"}],'
-            b'"media":[{"id":"m1"}]}}\n'
-        )
-        assert state["includes"] == {"users": 1, "tweets": 2, "media": 1}
-
-    def test_unknown_include_keys_are_bounded_and_known_keys_continue(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-
-        for index in range(70):
-            parse(
-                b'{"data":{"id":"'
-                + str(index).encode()
-                + b'"},"includes":{"future_'
-                + str(index).encode()
-                + b'":[{"id":"u"}]}}\n'
-            )
-        parse(b'{"data":{"id":"known"},"includes":{"users":[{"id":"user"}]}}\n')
-
-        unknown_keys = [key for key in state["includes"] if key.startswith("future_")]
-        assert len(unknown_keys) == 64
-        assert state["unknown_includes_overflow_count"] == 6
-        assert state["includes"]["users"] == 1
-        assert state["data_count"] == 71
-        assert state["lines_failed"] == 0
-
-    def test_unsafe_unknown_include_keys_overflow(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        overlong_key = b"x" * 92
-
-        parse(
-            b'{"data":{"id":"1"},"includes":{"'
-            + overlong_key
-            + b'":[{"id":"long"}],"bad/key":[{"id":"one"},{"id":"two"}],'
-            b'"__overflow__":[{"id":"reserved"}]}}\n'
-        )
-
-        assert state["includes"] == {}
-        assert state["unknown_includes_overflow_count"] == 4
-        assert state["data_count"] == 1
-        assert state["lines_failed"] == 0
-
-    def test_data_array_not_counted(self, real_flow):
-        """Line where top-level ``data`` is an array (not a dict) contributes 0 to data_count."""
-        parse, state = self._stream_parser(real_flow)
-        parse(b'{"data":[1,2,3]}\n')
-        assert state["data_count"] == 0
-        assert state["lines_parsed"] == 1
-
-    def test_non_dict_top_level_skipped(self, real_flow):
-        parse, state = self._stream_parser(real_flow)
-        parse(b'"some string"\n')
-        parse(b"42\n")
-        parse(b'{"data":{"id":"1"}}\n')
-        assert state["lines_parsed"] == 3
-        assert state["data_count"] == 1
-
-
-class TestXJsonFinalize:
-    """Tests for finalizing non-streaming X JSON parser state."""
-
-    def _billable_x_json_flow(self, real_flow):
-        return make_x_response_flow(real_flow)
-
-    def test_no_registered_x_json_finalizer_is_noop(self, real_flow):
-        flow = real_flow(with_response=False, host="api.x.com", path="/2/tweets")
-
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert metadata_keys.X_JSON_STATE not in flow.metadata
-
-    def test_finalizes_successful_x_json_state(self, real_flow):
-        flow = self._billable_x_json_flow(real_flow)
-        mitm_addon.responseheaders(flow)
-        assert "connector_response_finish" in flow.metadata
-
-        response_stream(flow)(b'{"data":[{"id":"1"}]}')
-        response_streaming.finalize_connector_response_state(flow)
-
-        state = dict(flow.metadata[metadata_keys.X_JSON_STATE])
-        assert "connector_response_finish" not in flow.metadata
-        assert state["body_parsed"] is True
-        assert state["response_data_count"] == 1
-        assert "parse_error" not in state
-
-        response_streaming.finalize_connector_response_state(flow)
-        assert flow.metadata[metadata_keys.X_JSON_STATE] == state
-
-    def test_forensic_buffer_truncation_does_not_stop_x_json_parser(self, real_flow):
-        flow = make_x_response_flow(
-            real_flow,
-            path="/2/users/by",
-            original_url="https://api.x.com/2/users/by?ids=1,2,3",
-        )
-        flow.metadata[metadata_keys.CAPTURE_BODY] = True
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        callback(b'{"data":[{"id":"1","text":"')
-        callback(b"x" * (200 * 1024))
-        callback(
-            b'"}],"includes":{"users":[{"id":"u1"}]},'
-            b'"meta":{"result_count":1,"total_tweet_count":2}}'
-        )
-        response_streaming.finalize_connector_response_state(flow)
-
-        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
-        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
-        assert metadata_keys.X_NDJSON_STATE not in flow.metadata
-        state = flow.metadata[metadata_keys.X_JSON_STATE]
-        assert state["body_parsed"] is True
-        assert state["response_data_count"] == 1
-        assert state["response_includes"] == {"users": 1}
-        assert state["response_result_count"] == 1
-        assert state["response_total_tweet_count"] == 2
-
-    def test_decompresses_gzip_x_json_before_parsing(self, real_flow):
-        body = (
-            b'{"data":[{"id":"1"},{"id":"2"}],'
-            b'"includes":{"users":[{"id":"u1"}]},'
-            b'"meta":{"result_count":2,"total_tweet_count":3}}'
-        )
-        flow = make_x_response_flow(real_flow, content_encoding="gzip")
-
-        mitm_addon.responseheaders(flow)
-
-        response_stream(flow)(gzip.compress(body))
-        response_streaming.finalize_connector_response_state(flow)
-        state = flow.metadata[metadata_keys.X_JSON_STATE]
-        assert state["response_data_count"] == 2
-        assert state["response_includes"] == {"users": 1}
-        assert state["response_result_count"] == 2
-        assert state["response_total_tweet_count"] == 3
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_x_json_feeds_later_members(self, real_flow, encoding):
-        body = (
-            b'{"data":[{"id":"1"},{"id":"2"}],'
-            b'"includes":{"users":[{"id":"u1"}]},'
-            b'"meta":{"result_count":2,"total_tweet_count":3}}'
-        )
-        if encoding == "gzip":
-            compressed = gzip.compress(b"") + gzip.compress(body)
-        else:
-            compressed = zlib.compress(b"") + zlib.compress(body)
-        flow = make_x_response_flow(real_flow, content_encoding=encoding)
-
-        mitm_addon.responseheaders(flow)
-        response_stream(flow)(compressed)
-        response_streaming.finalize_connector_response_state(flow)
-
-        state = flow.metadata[metadata_keys.X_JSON_STATE]
-        assert state["body_parsed"] is True
-        assert state["response_data_count"] == 2
-        assert state["response_includes"] == {"users": 1}
-        assert state["response_result_count"] == 2
-        assert state["response_total_tweet_count"] == 3
-
-    def test_finalizes_x_json_parse_error(self, real_flow):
-        flow = self._billable_x_json_flow(real_flow)
-        mitm_addon.responseheaders(flow)
-        assert "connector_response_finish" in flow.metadata
-
-        response_stream(flow)(b'{"data":[1')
-        response_streaming.finalize_connector_response_state(flow)
-
-        state = dict(flow.metadata[metadata_keys.X_JSON_STATE])
-        assert "connector_response_finish" not in flow.metadata
-        assert state["body_parsed"] is False
-        assert isinstance(state["parse_error"], str)
-        assert state["parse_error"]
-
-        response_streaming.finalize_connector_response_state(flow)
-        assert flow.metadata[metadata_keys.X_JSON_STATE] == state
-
-    def test_finalizes_non_object_x_json_without_parse_error(self, real_flow):
-        flow = self._billable_x_json_flow(real_flow)
-        mitm_addon.responseheaders(flow)
-        assert "connector_response_finish" in flow.metadata
-
-        response_stream(flow)(b"[1,2,3]")
-        response_streaming.finalize_connector_response_state(flow)
-
-        state = flow.metadata[metadata_keys.X_JSON_STATE]
-        assert "connector_response_finish" not in flow.metadata
-        assert state["body_parsed"] is False
-        assert "parse_error" not in state
-
-
-class TestResponseHeadersModelJsonParser:
-    """Tests for model-provider JSON parser setup in responseheaders()."""
-
-    def test_brotli_model_json_skips_incremental_parser(self, real_flow, mitm_ctx):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "application/json", "content-encoding": "br"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        with mitm_ctx() as log:
-            mitm_addon.responseheaders(flow)
-
-        assert callable(response_stream(flow))
-        assert "model_json_usage_finish" not in flow.metadata
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert any(
-            "Streaming decompression skipped: brotli streaming output cannot be bounded"
-            in call.args[0]
-            for call in log.debug.call_args_list
-        )
-
-
-class TestResponseHeadersSseParser:
-    """Tests for SSE parser setup in responseheaders()."""
-
-    def test_sets_up_sse_parser_for_model_provider(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        assert isinstance(flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE], dict)
-        assert "model_sse_usage_finish" in flow.metadata
-        assert "model_json_usage_finish" not in flow.metadata
-        # Feed SSE data through the callback
-        callback = response_stream(flow)
-        callback(
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":'
-            b'{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":42}}}\n\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "claude-sonnet-4-6"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.input"] == 42
-
-    def test_sets_up_sse_parser_with_case_insensitive_content_type(self, real_flow):
-        flow = real_flow(with_response=False, host="api.openai.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "Text/Event-Stream"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:openai-api-key"
-        flow.metadata[metadata_keys.CLI_AGENT_TYPE] = "codex"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-5.5"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        callback = response_stream(flow)
-        callback(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.5",'
-            b'"usage":{"output_tokens":5}}}\n\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "gpt-5.5"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.output"] == 5
-
-    def test_finalizes_sse_parser_for_trailing_event_without_blank_line(self, real_flow):
-        flow = real_flow(with_response=False, host="api.openai.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "text/event-stream"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:openai-api-key"
-        flow.metadata[metadata_keys.CLI_AGENT_TYPE] = "codex"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-5.5"
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        callback(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.5",'
-            b'"usage":{"output_tokens":7}}}\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-
-        response_streaming.finalize_model_sse_usage(flow)
-
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "gpt-5.5"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.output"] == 7
-
-    def test_sets_up_openai_sse_parser_for_openai_model_provider(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.openai.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "text/event-stream"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:openai-api-key"
-        flow.metadata[metadata_keys.CLI_AGENT_TYPE] = "codex"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-5.5"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        callback = response_stream(flow)
-        callback(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":42,'
-            b'"input_tokens_details":{"cached_tokens":12}}}}\n\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "gpt-5.5"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.input"] == 30
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.cache_read"] == 12
-
-    def test_codex_oauth_model_provider_uses_openai_sse_parser(self, real_flow):
-        flow = real_flow(with_response=False, host="chatgpt.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "text/event-stream"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:codex-oauth-token"
-        flow.metadata[metadata_keys.CLI_AGENT_TYPE] = "codex"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-5.5"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        callback = response_stream(flow)
-        callback(
-            b"event: response.completed\n"
-            b'data: {"response":{"model":"gpt-5.5",'
-            b'"usage":{"input_tokens":42,'
-            b'"input_tokens_details":{"cached_tokens":12}}}}\n\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "gpt-5.5"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.input"] == 30
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.cache_read"] == 12
-
-    @pytest.mark.parametrize("cli_agent_type", [None, ""])
-    def test_default_cli_agent_type_uses_anthropic_sse_parser(self, real_flow, cli_agent_type):
-        flow = real_flow(with_response=False, host="chatgpt.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": "text/event-stream"}),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:codex-oauth-token"
-        if cli_agent_type is not None:
-            flow.metadata[metadata_keys.CLI_AGENT_TYPE] = cli_agent_type
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        callback = response_stream(flow)
-        callback(
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":'
-            b'{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":42}}}\n\n'
-        )
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "claude-sonnet-4-6"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.input"] == 42
-
-    def test_decompresses_gzip_sse_before_parsing(self, real_flow, headers):
-        """Compressed SSE streams must be decompressed before usage extraction."""
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map(
-                {
-                    "content-type": "text/event-stream; charset=utf-8",
-                    "content-encoding": "gzip",
-                }
-            ),
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        callback = response_stream(flow)
-        plaintext = (
-            b"event: message_start\n"
-            b'data: {"type":"message_start","message":'
-            b'{"model":"claude-sonnet-4-6",'
-            b'"usage":{"input_tokens":99}}}\n\n'
-        )
-        compressed = gzip.compress(plaintext)
-        # Callback returns original compressed bytes to client
-        result = callback(compressed)
-        assert result == compressed
-        # But parser receives decompressed data
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["model"] == "claude-sonnet-4-6"
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE]["tokens.input"] == 99
-
-    def test_no_sse_parser_for_non_model_provider(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.github.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "github"
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert "model_json_usage_finish" not in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-
-    def test_no_sse_parser_for_non_sse_response(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        assert "model_json_usage_finish" in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-
-    def test_no_sse_parser_without_model_usage_provider(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-
-    def test_sets_up_sse_parser_for_non_billable_observable_model_provider(
-        self, real_flow, headers
-    ):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = False
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE in flow.metadata
-        assert isinstance(flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE], dict)
-        assert "model_sse_usage_finish" in flow.metadata
-
-    def test_no_sse_parser_without_firewall_name(self, real_flow, headers):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "text/event-stream"})
-        )
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-        # Model metadata alone must not classify the flow as a model provider.
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert "model_json_usage_finish" not in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-
-    @pytest.mark.parametrize("firewall_name", [None, 42])
-    def test_malformed_firewall_name_skips_usage_parsers(self, real_flow, firewall_name):
-        flow = make_x_response_flow(
-            real_flow,
-            path="/2/tweets/search/stream",
-            firewall_name=firewall_name,
-        )
-
-        mitm_addon.responseheaders(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE not in flow.metadata
-        assert "model_json_usage_finish" not in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-        assert "connector_response_finish" not in flow.metadata
-        assert metadata_keys.X_NDJSON_STATE not in flow.metadata
-
-
-class TestReleaseResponseStreamState:
-    """Tests for direct response streaming cleanup."""
-
-    def test_preserves_externally_replaced_stream_callback(self, real_flow):
-        flow = real_flow(with_response=False, host="api.example.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        def external_stream(chunk):
-            return chunk
-
-        mitm_addon.responseheaders(flow)
-        assert callable(response_stream(flow))
-
-        flow.response.stream = external_stream
-
-        response_streaming.release_response_stream_state(flow)
-
-        assert flow.response.stream is external_stream
-        assert "_vm0_response_stream_callback" not in flow.metadata
-        assert metadata_keys.RESPONSE_STREAM_STATE not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    @pytest.mark.parametrize(
-        (
-            "host",
-            "path",
-            "response_content_type",
-            "metadata",
-            "finish_key",
-            "reports_on_interruption",
-        ),
-        [
-            pytest.param(
-                "api.anthropic.com",
-                "/v1/messages",
-                "application/json",
-                {
-                    metadata_keys.FIREWALL_NAME: "model-provider:anthropic-api-key",
-                    metadata_keys.FIREWALL_BILLABLE: True,
-                    metadata_keys.MODEL_USAGE_PROVIDER: "claude-sonnet-4-6",
-                },
-                "model_json_usage_finish",
-                False,
-                id="model-json",
-            ),
-            pytest.param(
-                "api.openai.com",
-                "/v1/responses",
-                "text/event-stream",
-                {
-                    metadata_keys.FIREWALL_NAME: "model-provider:openai-api-key",
-                    metadata_keys.CLI_AGENT_TYPE: "codex",
-                    metadata_keys.FIREWALL_BILLABLE: True,
-                    metadata_keys.MODEL_USAGE_PROVIDER: "gpt-5.5",
-                },
-                "model_sse_usage_finish",
-                False,
-                id="model-sse",
-            ),
-            pytest.param(
-                "api.x.com",
-                "/2/tweets",
-                "application/json",
-                {
-                    metadata_keys.FIREWALL_NAME: "x",
-                    metadata_keys.FIREWALL_BILLABLE: True,
-                    metadata_keys.ORIGINAL_URL: "https://api.x.com/2/tweets",
-                },
-                "connector_response_finish",
-                False,
-                id="x-json",
-            ),
-            pytest.param(
-                "api.x.com",
-                "/2/tweets/search/stream",
-                "application/json",
-                {
-                    metadata_keys.FIREWALL_NAME: "x",
-                    metadata_keys.FIREWALL_BILLABLE: True,
-                    metadata_keys.ORIGINAL_URL: "https://api.x.com/2/tweets/search/stream",
-                },
-                "connector_response_finish",
-                True,
-                id="x-ndjson",
-            ),
-        ],
-    )
-    def test_removes_unfinalized_parser_finish_callback(
-        self,
-        real_flow,
-        host,
-        path,
-        response_content_type,
-        metadata,
-        finish_key,
-        reports_on_interruption,
-    ):
-        flow = real_flow(with_response=False, host=host, path=path)
-        flow.metadata.update(metadata)
-        flow.response = tutils.tresp(
-            status_code=200,
-            headers=header_map({"content-type": response_content_type}),
-        )
-
-        mitm_addon.responseheaders(flow)
-        assert finish_key in flow.metadata
-        assert (
-            "connector_response_report_on_interruption" in flow.metadata
-        ) is reports_on_interruption
-
-        response_streaming.release_response_stream_state(flow)
-
-        assert finish_key not in flow.metadata
-        assert "connector_response_report_on_interruption" not in flow.metadata
-        assert metadata_keys.RESPONSE_STREAM_STATE not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-        assert flow.response.stream is False
-
-    def test_unconfigured_flow_is_noop(self, real_flow):
-        flow = real_flow(with_response=True, host="api.example.com")
-
-        def external_stream(chunk):
-            return chunk
-
-        assert flow.response is not None
-        flow.response.stream = external_stream
-
-        response_streaming.release_response_stream_state(flow)
-
-        assert flow.response.stream is external_stream
-
-    def test_configured_release_is_idempotent(self, real_flow):
-        flow = real_flow(with_response=False, host="api.example.com")
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        mitm_addon.responseheaders(flow)
-        assert callable(response_stream(flow))
-
-        response_streaming.release_response_stream_state(flow)
-        response_streaming.release_response_stream_state(flow)
-
-        assert flow.response.stream is False
-        assert "_vm0_response_stream_callback" not in flow.metadata
-        assert metadata_keys.RESPONSE_STREAM_STATE not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-    def test_release_after_response_is_removed_still_drops_metadata(self, real_flow):
-        flow = real_flow(with_response=False, host="api.anthropic.com")
-        flow.metadata[metadata_keys.FIREWALL_NAME] = "model-provider:anthropic-api-key"
-        flow.metadata[metadata_keys.FIREWALL_BILLABLE] = True
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "claude-sonnet-4-6"
-        flow.response = tutils.tresp(
-            status_code=200, headers=header_map({"content-type": "application/json"})
-        )
-
-        mitm_addon.responseheaders(flow)
-        assert callable(response_stream(flow))
-        assert "model_json_usage_finish" in flow.metadata
-
-        flow.response = None
-
-        response_streaming.release_response_stream_state(flow)
-
-        assert flow.response is None
-        assert "_vm0_response_stream_callback" not in flow.metadata
-        assert metadata_keys.RESPONSE_STREAM_STATE not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-        assert "model_json_usage_finish" not in flow.metadata
