+++ crates/runner/mitm-addon/tests/test_x_response_parsers.py
@@ -1,0 +1,433 @@
+"""X connector response parser integration tests."""
+
+import gzip
+import zlib
+
+import pytest
+
+import flow_metadata_keys as metadata_keys
+import mitm_addon
+import response_streaming
+from body_limits import LARGE_RESPONSE_DECOMPRESS_LIMIT, STREAM_BUFFER_LIMIT
+from tests.flow_helpers import response_stream
+from tests.x_flow_helpers import make_x_response_flow
+
+_OVERSIZED_NDJSON_LINE_BYTES = LARGE_RESPONSE_DECOMPRESS_LIMIT + 1024
+
+
+class TestNdjsonExtractor:
+    """Tests for X NDJSON extraction through responseheaders (issue #9534)."""
+
+    def _stream_flow(self, real_flow, *, capture_body=False):
+        flow = make_x_response_flow(real_flow, path="/2/tweets/search/stream")
+        if capture_body:
+            flow.metadata[metadata_keys.CAPTURE_BODY] = True
+        mitm_addon.responseheaders(flow)
+        return flow
+
+    def _stream_parser(self, real_flow):
+        flow = self._stream_flow(real_flow)
+        return response_stream(flow), flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+    def test_single_line(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n')
+        assert state["data_count"] == 1
+        assert state["includes"] == {"users": 1}
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 0
+
+    def test_multiple_lines_aggregate_counts(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n')
+        parse(b'{"data":{"id":"2"},"includes":{"users":[{"id":"u2"},{"id":"u3"}]}}\n')
+        assert state["data_count"] == 2
+        assert state["includes"] == {"users": 3}
+        assert state["lines_parsed"] == 2
+
+    def test_forensic_buffer_truncation_does_not_stop_parser(self, real_flow):
+        flow = self._stream_flow(real_flow, capture_body=True)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+        parse(b'{"data":{"id":"1"}}\n' + b"x" * (200 * 1024))
+
+        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
+        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
+        assert state["data_count"] == 1
+        assert "connector_response_finish" in flow.metadata
+
+    def test_decompresses_gzip_stream_before_parsing(self, real_flow):
+        ndjson_body = (
+            b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}\n'
+            b'{"data":{"id":"2"},"includes":{"users":[{"id":"u2"}]}}\n'
+            b'{"data":{"id":"3"},"includes":{"users":[{"id":"u3"}]}}\n'
+        )
+        compressed = gzip.compress(ndjson_body)
+        flow = make_x_response_flow(
+            real_flow,
+            path="/2/tweets/search/stream",
+            content_encoding="gzip",
+        )
+
+        mitm_addon.responseheaders(flow)
+
+        mid = len(compressed) // 2
+        response_stream(flow)(compressed[:mid])
+        response_stream(flow)(compressed[mid:])
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+        assert state["data_count"] == 3
+        assert state["includes"] == {"users": 3}
+        assert "connector_response_finish" in flow.metadata
+
+    def test_chunked_line_split_mid_json(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"},"include')
+        parse(b's":{"users":[{"id":"u1"}]}}\n')
+        assert state["data_count"] == 1
+        assert state["includes"] == {"users": 1}
+        assert state["lines_parsed"] == 1
+
+    def test_keep_alive_blank_lines(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b"\n\n")
+        parse(b'{"data":{"id":"1"}}\n')
+        parse(b"\n")
+        parse(b'{"data":{"id":"2"}}\n')
+        assert state["data_count"] == 2
+        assert state["lines_parsed"] == 2
+
+    def test_crlf_line_endings(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"}}\r\n{"data":{"id":"2"}}\r\n')
+        assert state["data_count"] == 2
+        assert state["lines_parsed"] == 2
+
+    def test_crlf_line_ending_split_across_chunks(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"}}\r')
+        parse(b'\n{"data":{"id":"2"}}\r\n')
+        assert state["data_count"] == 2
+        assert state["lines_parsed"] == 2
+        assert state["lines_failed"] == 0
+
+    def test_malformed_line_increments_failures(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"}}\n')
+        parse(b"not json at all\n")
+        parse(b'{"data":{"id":"2"}}\n')
+        assert state["data_count"] == 2
+        assert state["lines_parsed"] == 2
+        assert state["lines_failed"] == 1
+
+    def test_invalid_utf8_line_increments_failures_and_continues(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'\x80{"data":{"id":"bad"}}\n{"data":{"id":"after"}}\n')
+
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 1
+
+    def test_truncated_trailing_line_not_counted(self, real_flow):
+        """Connection drops mid-line — partial trailing line stays in buf, not counted."""
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":{"id":"1"}}\n{"data":{"id":"2"}')  # no trailing \n
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+
+    def test_complete_trailing_line_without_newline_counted_on_finish(self, real_flow):
+        flow = self._stream_flow(real_flow)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+        parse(b'{"data":{"id":"1"},"includes":{"users":[{"id":"u1"}]}}')
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert state["data_count"] == 1
+        assert state["includes"] == {"users": 1}
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 0
+
+    def test_incomplete_trailing_line_without_newline_fails_on_finish(self, real_flow):
+        flow = self._stream_flow(real_flow)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+        parse(b'{"data":{"id":"1"}}\n{"data":{"id":"2"}')
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 1
+
+    @pytest.mark.parametrize("tail", [b"", b"\r", b"\r\r"])
+    def test_blank_trailing_line_ignored_on_finish(self, real_flow, tail):
+        flow = self._stream_flow(real_flow)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+        parse(b'{"data":{"id":"1"}}\n' + tail)
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 0
+
+    def test_trailing_line_finish_is_idempotent(self, real_flow):
+        flow = self._stream_flow(real_flow)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+
+        parse(b'{"data":{"id":"1"}}')
+        response_streaming.finalize_connector_response_state(flow)
+        finalized = dict(state)
+        finalized["includes"] = dict(state["includes"])
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert state == finalized
+
+    def test_oversized_line_tail_not_counted_on_finish(self, real_flow):
+        flow = self._stream_flow(real_flow)
+        parse = response_stream(flow)
+        state = flow.metadata[metadata_keys.X_NDJSON_STATE]
+        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
+
+        parse(big)
+        parse(b'{"data":{"id":"tail"}}')
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert state["data_count"] == 0
+        assert state["lines_parsed"] == 0
+        assert state["lines_failed"] == 1
+
+    def test_empty_chunks_safe(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b"")
+        parse(b'{"data":{"id":"1"}}\n')
+        parse(b"")
+        assert state["data_count"] == 1
+
+    def test_oversized_line_dropped(self, real_flow):
+        """Oversized line is dropped; subsequent lines parse normally."""
+        parse, state = self._stream_parser(real_flow)
+        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
+        parse(big)
+        parse(b"\n")
+        parse(b'{"data":{"id":"after"}}\n')
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 1
+
+    def test_oversized_line_discards_until_newline(self, real_flow):
+        """A valid-looking tail of an overlong line must not be counted as its own row."""
+        parse, state = self._stream_parser(real_flow)
+        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
+        parse(big)
+        parse(b'{"data":{"id":"tail"}}\n')
+        parse(b'{"data":{"id":"next"}}\n')
+
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 1
+
+    def test_oversized_line_with_newline_continues_in_same_chunk(self, real_flow):
+        """Dropping an overlong row should not discard valid later rows in the same chunk."""
+        parse, state = self._stream_parser(real_flow)
+        big = b"x" * _OVERSIZED_NDJSON_LINE_BYTES
+        parse(big + b'\n{"data":{"id":"after"}}\n')
+
+        assert state["data_count"] == 1
+        assert state["lines_parsed"] == 1
+        assert state["lines_failed"] == 1
+
+    def test_includes_multiple_keys(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(
+            b'{"data":{"id":"1"},"includes":'
+            b'{"users":[{"id":"u1"}],'
+            b'"tweets":[{"id":"t1"},{"id":"t2"}],'
+            b'"media":[{"id":"m1"}]}}\n'
+        )
+        assert state["includes"] == {"users": 1, "tweets": 2, "media": 1}
+
+    def test_unknown_include_keys_are_bounded_and_known_keys_continue(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+
+        for index in range(70):
+            parse(
+                b'{"data":{"id":"'
+                + str(index).encode()
+                + b'"},"includes":{"future_'
+                + str(index).encode()
+                + b'":[{"id":"u"}]}}\n'
+            )
+        parse(b'{"data":{"id":"known"},"includes":{"users":[{"id":"user"}]}}\n')
+
+        unknown_keys = [key for key in state["includes"] if key.startswith("future_")]
+        assert len(unknown_keys) == 64
+        assert state["unknown_includes_overflow_count"] == 6
+        assert state["includes"]["users"] == 1
+        assert state["data_count"] == 71
+        assert state["lines_failed"] == 0
+
+    def test_unsafe_unknown_include_keys_overflow(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        overlong_key = b"x" * 92
+
+        parse(
+            b'{"data":{"id":"1"},"includes":{"'
+            + overlong_key
+            + b'":[{"id":"long"}],"bad/key":[{"id":"one"},{"id":"two"}],'
+            b'"__overflow__":[{"id":"reserved"}]}}\n'
+        )
+
+        assert state["includes"] == {}
+        assert state["unknown_includes_overflow_count"] == 4
+        assert state["data_count"] == 1
+        assert state["lines_failed"] == 0
+
+    def test_data_array_not_counted(self, real_flow):
+        """Line where top-level ``data`` is an array (not a dict) contributes 0 to data_count."""
+        parse, state = self._stream_parser(real_flow)
+        parse(b'{"data":[1,2,3]}\n')
+        assert state["data_count"] == 0
+        assert state["lines_parsed"] == 1
+
+    def test_non_dict_top_level_skipped(self, real_flow):
+        parse, state = self._stream_parser(real_flow)
+        parse(b'"some string"\n')
+        parse(b"42\n")
+        parse(b'{"data":{"id":"1"}}\n')
+        assert state["lines_parsed"] == 3
+        assert state["data_count"] == 1
+
+
+class TestXJsonFinalize:
+    """Tests for finalizing non-streaming X JSON parser state."""
+
+    def _billable_x_json_flow(self, real_flow):
+        return make_x_response_flow(real_flow)
+
+    def test_no_registered_x_json_finalizer_is_noop(self, real_flow):
+        flow = real_flow(with_response=False, host="api.x.com", path="/2/tweets")
+
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert metadata_keys.X_JSON_STATE not in flow.metadata
+
+    def test_finalizes_successful_x_json_state(self, real_flow):
+        flow = self._billable_x_json_flow(real_flow)
+        mitm_addon.responseheaders(flow)
+        assert "connector_response_finish" in flow.metadata
+
+        response_stream(flow)(b'{"data":[{"id":"1"}]}')
+        response_streaming.finalize_connector_response_state(flow)
+
+        state = dict(flow.metadata[metadata_keys.X_JSON_STATE])
+        assert "connector_response_finish" not in flow.metadata
+        assert state["body_parsed"] is True
+        assert state["response_data_count"] == 1
+        assert "parse_error" not in state
+
+        response_streaming.finalize_connector_response_state(flow)
+        assert flow.metadata[metadata_keys.X_JSON_STATE] == state
+
+    def test_forensic_buffer_truncation_does_not_stop_x_json_parser(self, real_flow):
+        flow = make_x_response_flow(
+            real_flow,
+            path="/2/users/by",
+            original_url="https://api.x.com/2/users/by?ids=1,2,3",
+        )
+        flow.metadata[metadata_keys.CAPTURE_BODY] = True
+        mitm_addon.responseheaders(flow)
+
+        callback = response_stream(flow)
+        callback(b'{"data":[{"id":"1","text":"')
+        callback(b"x" * (200 * 1024))
+        callback(
+            b'"}],"includes":{"users":[{"id":"u1"}]},'
+            b'"meta":{"result_count":1,"total_tweet_count":2}}'
+        )
+        response_streaming.finalize_connector_response_state(flow)
+
+        assert len(flow.metadata[metadata_keys.STREAM_BUFFER]) == STREAM_BUFFER_LIMIT
+        assert flow.metadata[metadata_keys.STREAM_BUFFER_STATE]["truncated"] is True
+        assert metadata_keys.X_NDJSON_STATE not in flow.metadata
+        state = flow.metadata[metadata_keys.X_JSON_STATE]
+        assert state["body_parsed"] is True
+        assert state["response_data_count"] == 1
+        assert state["response_includes"] == {"users": 1}
+        assert state["response_result_count"] == 1
+        assert state["response_total_tweet_count"] == 2
+
+    def test_decompresses_gzip_x_json_before_parsing(self, real_flow):
+        body = (
+            b'{"data":[{"id":"1"},{"id":"2"}],'
+            b'"includes":{"users":[{"id":"u1"}]},'
+            b'"meta":{"result_count":2,"total_tweet_count":3}}'
+        )
+        flow = make_x_response_flow(real_flow, content_encoding="gzip")
+
+        mitm_addon.responseheaders(flow)
+
+        response_stream(flow)(gzip.compress(body))
+        response_streaming.finalize_connector_response_state(flow)
+        state = flow.metadata[metadata_keys.X_JSON_STATE]
+        assert state["response_data_count"] == 2
+        assert state["response_includes"] == {"users": 1}
+        assert state["response_result_count"] == 2
+        assert state["response_total_tweet_count"] == 3
+
+    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
+    def test_concatenated_zlib_x_json_feeds_later_members(self, real_flow, encoding):
+        body = (
+            b'{"data":[{"id":"1"},{"id":"2"}],'
+            b'"includes":{"users":[{"id":"u1"}]},'
+            b'"meta":{"result_count":2,"total_tweet_count":3}}'
+        )
+        if encoding == "gzip":
+            compressed = gzip.compress(b"") + gzip.compress(body)
+        else:
+            compressed = zlib.compress(b"") + zlib.compress(body)
+        flow = make_x_response_flow(real_flow, content_encoding=encoding)
+
+        mitm_addon.responseheaders(flow)
+        response_stream(flow)(compressed)
+        response_streaming.finalize_connector_response_state(flow)
+
+        state = flow.metadata[metadata_keys.X_JSON_STATE]
+        assert state["body_parsed"] is True
+        assert state["response_data_count"] == 2
+        assert state["response_includes"] == {"users": 1}
+        assert state["response_result_count"] == 2
+        assert state["response_total_tweet_count"] == 3
+
+    def test_finalizes_x_json_parse_error(self, real_flow):
+        flow = self._billable_x_json_flow(real_flow)
+        mitm_addon.responseheaders(flow)
+        assert "connector_response_finish" in flow.metadata
+
+        response_stream(flow)(b'{"data":[1')
+        response_streaming.finalize_connector_response_state(flow)
+
+        state = dict(flow.metadata[metadata_keys.X_JSON_STATE])
+        assert "connector_response_finish" not in flow.metadata
+        assert state["body_parsed"] is False
+        assert isinstance(state["parse_error"], str)
+        assert state["parse_error"]
+
+        response_streaming.finalize_connector_response_state(flow)
+        assert flow.metadata[metadata_keys.X_JSON_STATE] == state
+
+    def test_finalizes_non_object_x_json_without_parse_error(self, real_flow):
+        flow = self._billable_x_json_flow(real_flow)
+        mitm_addon.responseheaders(flow)
+        assert "connector_response_finish" in flow.metadata
+
+        response_stream(flow)(b"[1,2,3]")
+        response_streaming.finalize_connector_response_state(flow)
+
+        state = flow.metadata[metadata_keys.X_JSON_STATE]
+        assert "connector_response_finish" not in flow.metadata
+        assert state["body_parsed"] is False
+        assert "parse_error" not in state
